# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document APIs",
        "contact": {}
    },
    "paths": {
...
        "/cms/v1/strategies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Strategies",
                "parameters": [
                    {
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "DataStatusEnable",
                            "DataStatusDisable"
                        ],
                        "name": "data_status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at.desc",
                        "name": "sorts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.StrategyCmsDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Create Strategy",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.StrategyCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.StrategyCmsDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    }
                }
            }
        },
        "/cms/v1/strategies/{strategy_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Update Strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "strategy_id",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.StrategyCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.StrategyCmsDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ecode.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.StrategyAttributeInfo": {
            "type": "object",
            "required": [
                "key",
                "title",
                "type"
            ],
            "properties": {
                "default": {},
                "key": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "values": {}
            }
        },
        "db.StrategyCmsData": {
            "type": "object",
            "required": [
                "attributes",
                "data_status",
                "description",
                "filename",
                "name"
            ],
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.StrategyAttributeInfo"
                    }
                },
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "description": {
                    "type": "string",
                    "example": "Description"
                },
                "filename": {
                    "type": "string",
                    "example": "strategy_1.py"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                }
            }
        },
        "db.StrategyCmsDto": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.StrategyAttributeInfo"
                    }
                },
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "description": {
                    "type": "string",
                    "example": "Description"
                },
                "filename": {
                    "type": "string",
                    "example": "strategy_1.py"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "strategy_id": {
                    "type": "string",
                    "example": "671db9eca1f1b1bdbf3d4627"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "editor"
                }
            }
        },
...
        "ecode.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "error_description": {
                    "type": "string"
                }
            }
        },
...
        "enum.DataStatus": {
            "type": "string",
            "enum": [
                "enable",
                "disable"
            ],
            "x-enum-varnames": [
                "DataStatusEnable",
                "DataStatusDisable"
            ]
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Header"
        }
    }
}